% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateStudyAnalysesDetails.R
\name{createAnalysesDetails}
\alias{createAnalysesDetails}
\title{Create the analyses details}
\usage{
createAnalysesDetails(connectionDetails, cdmDatabaseSchema, workFolder)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{workFolder}{Name of local folder to place results; make sure to use forward slashes
(/)}
}
\description{
Create the analyses details
}
\details{
This function creates files specifying the analyses that will be performed.
}
