% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FiguresAndTablesForPaper.R
\name{createFiguresAndTables}
\alias{createFiguresAndTables}
\title{Generate diagnostics}
\usage{
createFiguresAndTables(outputFolder, connectionDetails, cohortDatabaseSchema,
  cohortTable, oracleTempSchema = oracleTempSchema)
}
\arguments{
\item{outputFolder}{Name of local folder where the results were generated; make sure to use forward slashes
(/). Do not use a folder on a network drive since this greatly impacts
performance.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cohortDatabaseSchema}{Schema name where intermediate data can be stored. You will need to have
write priviliges in this schema. Note that for SQL Server, this should
include both the database and schema name, for example 'cdm_data.dbo'.}

\item{cohortTable}{The name of the table that will be created in the work database schema.
This table will hold the exposure and outcome cohorts used in this
study.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
priviliges for storing temporary tables.}
}
\description{
Generate diagnostics
}
\details{
This function generates figures and tables.
}
