% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateTablesAndFigures.R
\name{createTableAndFigures}
\alias{createTableAndFigures}
\title{Create tables and figures}
\usage{
createTableAndFigures(exportFolder)
}
\arguments{
\item{exportFolder}{The path to the export folder containing the results.}
}
\description{
Create tables and figures
}
\details{
Creates tables and figures for viewing and interpreting the results. Requires that the
\code{\link{execute}} function has completed first.
}
