% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyApps.R
\name{prepareForEvidenceExplorer}
\alias{prepareForEvidenceExplorer}
\title{Prepare results for the Evidence Explorer Shiny app.}
\usage{
prepareForEvidenceExplorer(studyFolder)
}
\arguments{
\item{studyFolder}{The root folder containing the study results. The app expects each database to have a subfolder in this 
folder, containing the packaged results.}
}
\description{
Prepare results for the Evidence Explorer Shiny app.
}
